fix(android): 修复文件选择器闪退并完善导出与崩溃日志
- 文件选择改用宽松 MIME(*/*)并附带推荐类型，避免 ActivityNotFoundException
- pick/export launcher 内联到 HomePage，launch 加 runCatching 异常保护
- Uri 读取与产物导出移至 IO 线程，增加失败提示
- Application 增加全局未捕获异常日志(filesDir/nxshield/crash.log)
- 补齐编译期修复(material3-adaptive 依赖、ByteArray 写 ZIP、@Composable 标注等)

+++ android/app/src/main/java/com/nxshield/engine/Packer.kt
@@ -136,7 +136,7 @@ class Packer(
         val stats = Stats(
             classes = 0, dexCount = dexFiles.size,
             stringsProtected = stringsProtected, methodsExtracted = methodsExtracted,
-            assetsEncrypted = 0, elapsedMs = elapsed, outSize = outBytes.size,
+            assetsEncrypted = 0, elapsedMs = elapsed, outSize = outBytes.size.toLong(),
         )
         logger.log(jobId, "info", "task_done", mapOf(
             "elapsedMs" to elapsed, "extracted" to methodsExtracted,
